rewrite/reshape playwright tests
